+++ test/fixtures/char-class-empty-no-mbstring.php
@@ -1,7 +1,0 @@
-<?php
-/*
-ERROR GENERATING PARSER:
-
-Empty character class matching requires the `mbstring` PHP extension. However, the PHPeggy plugin option `mbstringAllowed` is set to `false`.
-
-*/