+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Released: TBD
 ### Breaking Changes
 
 - Node.js v20+ is now required
+- JavaScript API now requires explicitly option output = "source"
 - `mbstring` extension is required now. The `mbstringAllowed` option is removed.
 - Drop support for Internet Explorer